requirements.txt
bleach==3.1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dataclasses==0.6
docutils==0.16
enforce-pep8==0.0.3
idna==2.9
keyring==21.2.0
pkginfo==1.5.0.1
Pygments==2.6.1
pypans==0.0.7
PyYAML==5.3.1
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
termcolor==1.1.0
tqdm==4.45.0
twine==3.1.1
urequest==0.0.2
urllib3==1.25.8
uyaml==0.0.3
webencodings==0.5.1
flask==1.1.2
newspaper3k==0.2.8
gunicorn==20.0.4